requirements.txt
google-api-python-client==1.7.11
google-auth-httplib2
google-auth-oauthlib
O365==2.0.1
